Moved formatter factories to ./formatting/factories/
