+++ swingset/src/main/java/com/nqadmin/swingset/formatting/SSDateField.java
@@ -37,6 +37,7 @@
  ******************************************************************************/
 package com.nqadmin.swingset.formatting;
 
+import com.nqadmin.swingset.formatting.factories.SSDateFormatterFactory;
 
 // SSDateField.java
 //
