+++ swingset/src/main/java/com/nqadmin/swingset/formatting/SSIntegerField.java
@@ -39,6 +39,8 @@ package com.nqadmin.swingset.formatting;
 
 import javax.swing.SwingConstants;
 
+import com.nqadmin.swingset.formatting.factories.SSIntegerFormatterFactory;
+
 // SSIntegerField.java
 //
 // SwingSet - Open Toolkit For Making Swing Controls Database-Aware
