+++ swingset/src/main/java/com/nqadmin/swingset/formatting/factories/SSCurrencyFormatterFactory.java
@@ -35,7 +35,7 @@
  *   Man "Bee" Vo
  *   Ernie R. Rael
  ******************************************************************************/
-package com.nqadmin.swingset.formatting;
+package com.nqadmin.swingset.formatting.factories;
 
 import java.text.NumberFormat;
 import java.util.Locale;
