+++ swingset/src/main/java/com/nqadmin/swingset/formatting/factories/SSTimeFormatterFactory.java
@@ -35,7 +35,7 @@
  *   Man "Bee" Vo
  *   Ernie R. Rael
  ******************************************************************************/
-package com.nqadmin.swingset.formatting;
+package com.nqadmin.swingset.formatting.factories;
 
 import java.text.SimpleDateFormat;
